requirements.txt
# =============================================================================
# ARDT FMS - Python Requirements
# Version: 5.4
# =============================================================================

# Core Framework
Django>=5.1,<5.2

# Database
psycopg[binary]>=3.1

# Django Extensions
django-htmx>=1.17
django-widget-tweaks>=1.5
django-crispy-forms>=2.1
crispy-tailwind>=1.0
django-filter>=23.0
django-environ>=0.11

# Image Processing
Pillow>=10.0

# Barcode/QR Generation
python-barcode>=0.15
qrcode>=7.4

# Markdown (for Wiki)
markdown>=3.5
bleach>=6.0

# Production Server
whitenoise>=6.6
gunicorn>=21.0

# Development Tools
django-debug-toolbar>=4.2
django-extensions>=3.2

# Testing
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
factory-boy>=3.3

# Code Quality
black>=23.0
isort>=5.12
flake8>=6.0
